requirements.txt
# Web Framework and ASGI Server
fastapi==0.105.0          # Modern, fast web framework for building APIs
uvicorn==0.24.0           # Lightning-fast ASGI server implementation
websockets==12.0          # WebSocket server and client implementation for Python

# Database and ORM
sqlalchemy==2.0.23        # SQL toolkit and ORM for Python
aiosqlite==0.19.0         # Async SQLite support for FastAPI

# Data Validation and Serialization
pydantic==2.5.2          # Data validation using Python type annotations
pydantic-settings==2.1.0  # Settings management with Pydantic

# Template Engine and Static Files
jinja2==3.1.2            # Template engine for Python
python-multipart==0.0.6   # Handling multipart form data

# Data Processing
pandas==2.1.3            # Data manipulation and analysis library
pyyaml==6.0.1           # YAML parser and emitter
numpy==1.26.2           # Required by pandas, numerical computing library

# Environment and Configuration
python-dotenv==1.0.0     # Loading environment variables from .env files

# OpenAI Integration
openai==1.3.5            # Official OpenAI API client library

# Testing
pytest==7.4.3            # Testing framework
httpx==0.25.2            # HTTP client for testing FastAPI applications
pytest-asyncio==0.21.1   # Async support for pytest